Changements pour rendre les tests plus cohérents
